Solve error on non-ASCII commit messages (#282)
Solve UnicodeDecodeError that occurs when trying to make a commit who's
message contains non-ASCII characters (UnicodeDecodeError: 'ascii' codec
can't decode byte 0xc3 in position X: ordinal not in range (128)").

This error was described in issue #282.

The proposed solution involves indicating the exact encoding
(encoding='utf-8') when reading the file in use for writing the commit
message.